Added run script function
- Added System_run function (carried over from Je2D)
- Finished reworking precision
- Added deeper material control from lua

+++ Objects/Mesh.h
@@ -17,13 +17,13 @@ namespace FCE
 			void onRender();
 			std::string getOType();
 			
-			int load(std::string filename);
+			u8 load(std::string filename);
 			
 			scene::IAnimatedMesh* getMesh();
 			//scene::IAnimatedMeshSceneNode* getNode();
 			
 		private:
-			int id;
+			u32 id;
 			LOGGER* log;
 			
 			scene::IAnimatedMesh* mesh;
